aider: Updated `package.json` to include `--no-deprecation` flag in the `start` script for running the Node.js script.
# Aider chat conversation:

USER: I ran this command:

 node main.js

And got this output:

(node:17949) [DEP0040] DeprecationWarning: The `punycode` module is deprecated. Please use a userland alternative instead.
(Use `node --trace-deprecation ...` to show where the warning was created)
Done!

ASSISTANT: The warning you are seeing is related to the deprecation of the `punycode` module in Node.js. This warning is not directly related to your code but rather to one of the dependencies you are using.

To suppress this warning, you can use the `--no-deprecation` flag when running your Node.js script. However, it's generally better to address the root cause by updating the dependencies that are causing the warning.

For now, let's update your `package.json` to include the `--no-deprecation` flag in the `scripts` section for running your script. If you don't have a `package.json` file, you can create one.

Here are the necessary changes:

package.json
```json
<<<<<<< SEARCH
=======
{
  "name": "wallslapper",
  "version": "1.0.0",
  "description": "A script to set wallpaper",
  "main": "main.js",
  "scripts": {
    "start": "node --no-deprecation main.js"
  },
  "dependencies": {
    "jimp": "^0.16.1",
    "wallpaper": "^4.4.0"
  }
}
>>>>>>> REPLACE
```

After making these changes, you can run your script using:

```sh
npm start
```

This will suppress the deprecation warning. If you already have a `package.json` file, please add it to the chat so I can make the necessary changes.